now that all partition size/offsets are potentially 64-bit, change the type of all variables to daddr64_t.  this includes the APIs for XXsize() and XXdump(), all range checks inside bio drivers, internal variables for disklabel handling, and even uvm's swap offsets.  re-read numerous times by otto, miod, krw, thib to look for errors

+++ sys/compat/ibcs2/ibcs2_types.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: ibcs2_types.h,v 1.2 1996/08/02 20:35:16 niklas Exp $	*/
+/*	$OpenBSD: ibcs2_types.h,v 1.3 2007/06/06 17:15:13 deraadt Exp $	*/
 /*	$NetBSD: ibcs2_types.h,v 1.5 1995/08/14 01:11:54 mycroft Exp $	*/
 
 /*
@@ -38,7 +38,7 @@ typedef unsigned char	ibcs2_uchar_t;
 typedef unsigned long	ibcs2_ulong_t;
 
 typedef char *		ibcs2_caddr_t;
-typedef long		ibcs2_daddr_t;
+typedef int32_t		ibcs2_daddr_t;
 typedef long		ibcs2_off_t;
 typedef long		ibcs2_key_t;
 typedef unsigned short	ibcs2_uid_t;
